memory/pyproject.toml
[project]
name = "memory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argparse>=1.4.0",
    "bert-score>=0.3.13",
    "dspy>=2.6.27",
    "google-genai>=1.21.1",
    "nltk>=3.9.1",
    "numpy>=2.3.1",
    "pathlib>=1.0.1",
    "regex>=2024.11.6",
    "rouge>=1.0.1",
    "tqdm[asyncio]>=4.67.1",
]
